make help more specific and allow empty info
made it so sub command specific help is given if the help subcommands succeed or precede a valid subcommand.

I did not use include on argv since include cant handle both args being arrays, instead i convert them to strings and use .match

This means that it is possible for empty input to reach all methods, which is good since info is meant to show the logged in site if no argument  is given. I had to add emptiness checks for the other functions and create a pizza function though.

The help check also overrides info's ability to get the stats of the sites "help", "-h", and "--help" so I had to add an extra check